// src/app/_shared/constants/models/Product.ts
export interface Product{
    productId?: number;
    nameProduct? : string;
    description?: string;
    price?: number;
    releaseDate?: any;
    discontinuedDate?: any,
    rating?: number;
    categoryId?:number,
    supplierId?:number
}

// export interface Product{
//     productId?: number;
//     nameProduct? : string;
//     description?: string;
//     releaseDate?: any;
//     rating?: number;
//     price?: number;
// }
// export interface Product{
//     productId: number;
//     nameProduct : string;
//     description: string;
//     price: number;
//     releaseDate: any;
//     discontinuedDate: any,
//     rating: number;
//     categoryId:number,
//     supplierId:number
// }